requirements.txt
discord.py
psutil
ezgiphy
python-dotenv
